requirements.txt
# asttokens==3.0.0
# certifi==2025.11.12
# comm==0.2.3
# debugpy==1.8.17
# decorator==5.2.1
# dnspython==2.8.0
# executing==2.2.1
# ipykernel==6.30.1
# ipython==9.6.0
# ipython_pygments_lexers==1.1.1
# jedi==0.19.2
# joblib==1.5.2
# jupyter_client==8.6.3
# jupyter_core==5.8.1
# matplotlib-inline==0.1.7
# nest-asyncio==1.6.0
# -e git+https://github.com/Debcode2006/NetworkSecurity.git@f78e2b3d1ed70dc1bd50d9ef8b34f33d937a20c2#egg=NetworkSecurity
# numpy==2.3.5
# pandas==2.3.3
# parso==0.8.5
# platformdirs==4.4.0
# prompt_toolkit==3.0.52
# psutil==7.1.0
# pure_eval==0.2.3
# Pygments==2.19.2
# pymongo==4.15.5
# python-dateutil==2.9.0.post0
# python-dotenv==1.2.1
# pytz==2025.2
# pywin32==311
# pyzmq==27.1.0
# scikit-learn==1.7.2
# scipy==1.16.3
# setuptools==80.9.0
# six==1.17.0
# stack-data==0.6.3
# threadpoolctl==3.6.0
# tornado==6.5.2
# traitlets==5.14.3
# tzdata==2025.2
# wcwidth==0.2.14
# wheel==0.45.1
python-dotenv
pandas
numpy
pymongo
certifi
scikit-learn
pymongo[srv]
pyaml
dill

##-e .